rename model_naming.py to utils_model_naming.py
